pyproject.toml
[project]
name = "musubi-tuner"
version = "0.2.14"
description = "Musubi Tuner by kohya_ss"
readme = "README.md"
requires-python = ">=3.10, <3.13"
dependencies = [
    "accelerate==1.6.0",
    "av==14.0.1",
    "bitsandbytes",
    "diffusers==0.32.1",
    "einops==0.7.0",
    "huggingface-hub==0.34.3",
    "opencv-python==4.10.0.84",
    "pillow>=11.3.0",
    "safetensors==0.4.5",
    # "sageattention>=1.0.6",
    "toml==0.10.2",
    "tqdm==4.67.1",
    "transformers==4.54.1",
    "voluptuous==0.15.2",
    # Wan2.1
    "ftfy==6.3.1",
    "easydict==1.13",
    # FLUX.1 Kontext
    "sentencepiece==0.2.1",
]

[project.optional-dependencies]
cu124 = [
    "torch>=2.5.1",
    "torchvision>=0.20.1",
]
cu128 = [
    "torch>=2.7.1",
    "torchvision>=0.22.1",
]
rocm71 = [
    "torch>=2.8.0",
    "torchvision>=0.21.0",
]

[tool.uv]
conflicts = [
  [
    { extra = "cu124" },
    { extra = "cu128" },
    { extra = "rocm71" },
  ],
]

[dependency-groups]
dev = [
  "ascii-magic==2.3.0",
  "matplotlib==3.10.0",
  "tensorboard",
  "prompt-toolkit==3.0.51",
  "ruff>=0.12.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
torch = [
  { index = "pytorch-cu124", extra = "cu124" },
  { index = "pytorch-cu128", extra = "cu128" },
]
torchvision = [
  { index = "pytorch-cu124", extra = "cu124" },
  { index = "pytorch-cu128", extra = "cu128" },
]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.ruff]
line-length = 132
indent-width = 4
target-version = "py39"

lint.ignore = [
  "E402",  # module-import-not-at-top-of-file
  "E712",  # true-false-comparison
  "E721",  # type-comparison
  "E722",  # bare-except
  "E731",  # lambda-assignment
  "E741",  # ambiguous-variable-name
  "F601",  # multi-value-repeated-key-literal
  "F841",  # unused-variable
]

# Whether to enforce exclude and extend-exclude patterns, even for paths that are passed to Ruff explicitly
# This is used to ensure that certain files are always excluded via IDE formatters or external tool calls (CI, AI, etc.)
force-exclude = true

# Single-path or Relative patterns
# Extends the default exclude patterns documented at https://docs.astral.sh/ruff/settings/#exclude
extend-exclude = [
  "src/musubi_tuner/flux/flux_models.py",
  "src/musubi_tuner/frame_pack/hunyuan.py",
  "src/musubi_tuner/frame_pack/hunyuan_video_packed.py",
  "src/musubi_tuner/frame_pack/k_diffusion_hunyuan.py",
  "src/musubi_tuner/hunyuan_model/autoencoder_kl_causal_3d.py",
  "src/musubi_tuner/hunyuan_model/pipeline_hunyuan_video.py",
  "src/musubi_tuner/modules/scheduling_flow_match_discrete.py",
  "src/musubi_tuner/modules/unet_causal_3d_blocks.py",
  "src/musubi_tuner/qwen_image/qwen_image_autoencoder_kl.py",
  "src/musubi_tuner/qwen_image/qwen_image_model.py",
  "src/musubi_tuner/wan/configs/__init__.py",
  "src/musubi_tuner/wan/configs/shared_config.py",
  "src/musubi_tuner/wan/configs/wan_i2v_14B.py",
  "src/musubi_tuner/wan/configs/wan_i2v_A14B.py",
  "src/musubi_tuner/wan/configs/wan_t2v_14B.py",
  "src/musubi_tuner/wan/configs/wan_t2v_1_3B.py",
  "src/musubi_tuner/wan/configs/wan_t2v_A14B.py",
  "src/musubi_tuner/wan/modules/attention.py",
  "src/musubi_tuner/wan/modules/clip.py",
  "src/musubi_tuner/wan/modules/model.py",
  "src/musubi_tuner/wan/modules/t5.py",
  "src/musubi_tuner/wan/modules/tokenizers.py",
  "src/musubi_tuner/wan/modules/vae.py",
  "src/musubi_tuner/wan/modules/xlm_roberta.py",
  "src/musubi_tuner/wan/utils/fm_solvers.py",
  "src/musubi_tuner/wan/utils/fm_solvers_unipc.py",
  "src/musubi_tuner/wan/utils/utils.py",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"

[tool.ruff.lint.per-file-ignores]
# Ignore all rules for files with copyright headers from external sources
# https://docs.astral.sh/ruff/settings/#lint_per-file-ignores
# # Ignore `E402` (import violations) in all `__init__.py` files, and in `path/to/file.py`.
# "__init__.py" = ["E402"]
# "path/to/file.py" = ["E402"]
# # Ignore `D` rules everywhere except for the `src/` directory.
# "!src/**.py" = ["D"]
"src/musubi_tuner/flux/flux_models.py" = ["ALL"]
"src/musubi_tuner/frame_pack/hunyuan.py" = ["ALL"]
"src/musubi_tuner/frame_pack/hunyuan_video_packed.py" = ["ALL"]
"src/musubi_tuner/frame_pack/k_diffusion_hunyuan.py" = ["ALL"]
"src/musubi_tuner/hunyuan_model/autoencoder_kl_causal_3d.py" = ["ALL"]
"src/musubi_tuner/hunyuan_model/pipeline_hunyuan_video.py" = ["ALL"]
"src/musubi_tuner/modules/scheduling_flow_match_discrete.py" = ["ALL"]
"src/musubi_tuner/modules/unet_causal_3d_blocks.py" = ["ALL"]
"src/musubi_tuner/qwen_image/qwen_image_autoencoder_kl.py" = ["ALL"]
"src/musubi_tuner/qwen_image/qwen_image_model.py" = ["ALL"]
"src/musubi_tuner/wan/configs/__init__.py" = ["ALL"]
"src/musubi_tuner/wan/configs/shared_config.py" = ["ALL"]
"src/musubi_tuner/wan/configs/wan_i2v_14B.py" = ["ALL"]
"src/musubi_tuner/wan/configs/wan_i2v_A14B.py" = ["ALL"]
"src/musubi_tuner/wan/configs/wan_t2v_14B.py" = ["ALL"]
"src/musubi_tuner/wan/configs/wan_t2v_1_3B.py" = ["ALL"]
"src/musubi_tuner/wan/configs/wan_t2v_A14B.py" = ["ALL"]
"src/musubi_tuner/wan/modules/attention.py" = ["ALL"]
"src/musubi_tuner/wan/modules/clip.py" = ["ALL"]
"src/musubi_tuner/wan/modules/model.py" = ["ALL"]
"src/musubi_tuner/wan/modules/t5.py" = ["ALL"]
"src/musubi_tuner/wan/modules/tokenizers.py" = ["ALL"]
"src/musubi_tuner/wan/modules/vae.py" = ["ALL"]
"src/musubi_tuner/wan/modules/xlm_roberta.py" = ["ALL"]
"src/musubi_tuner/wan/utils/fm_solvers.py" = ["ALL"]
"src/musubi_tuner/wan/utils/fm_solvers_unipc.py" = ["ALL"]
"src/musubi_tuner/wan/utils/utils.py" = ["ALL"]
